+++ weblate/formats/exporters.py
@@ -34,6 +34,7 @@ from translate.storage.csvl10n import csvfile
 
 import weblate
 from weblate.formats.base import FileFormat
+from weblate.formats.external import XlsxFormat
 from weblate.utils.site import get_site_url
 
 # Map to remove control chars except newlines and tabs
@@ -53,6 +54,13 @@ def register_exporter(exporter):
 def get_exporter(name):
     """Return registered exporter"""
     return EXPORTERS[name]
+
+
+def list_exporters():
+    return [
+        {'name': x.name, 'verbose': x.verbose}
+        for x in sorted(EXPORTERS.values(), key=lambda x: x.name)
+    ]
 
 
 class BaseExporter(object):
@@ -148,7 +156,7 @@ class BaseExporter(object):
         )
 
         # Save to response
-        response.write(FileFormat.serialize(self.storage))
+        response.write(self.serialize())
 
         return response
 
@@ -290,12 +298,26 @@ class CSVExporter(BaseExporter):
     def string_filter(self, text):
         """Avoid Excel interpreting text as formula.
 
-        This is really bad idea implemented in Excel as this change leads
+        This is really bad idea, implemented in Excel, as this change leads
         to displaying additional ' in all other tools, but this seems to be
-        what most people are get used to. Hopefully these chars are not widely
-        used at first position of translatable strings, so the harm is not
-        that big.
+        what most people have gotten used to. Hopefully these chars are not widely
+        used at first position of translatable strings, so that harm is reduced.
         """
         if text and text[0] in ('=', '+', '-', '@', '|', '%'):
             return "'{0}'".format(text.replace('|', '\\|'))
         return text
+
+
+@register_exporter
+class XlsxExporter(BaseExporter):
+    name = 'xlsx'
+    content_type = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+    extension = 'xlsx'
+    verbose = _('Excel Open XML')
+
+    def get_storage(self):
+        return csvfile(fieldnames=self.fieldnames)
+
+    def serialize(self):
+        """Return storage content"""
+        return XlsxFormat.serialize(self.storage)