+++ weblate/trans/forms.py
@@ -1306,7 +1306,7 @@ class ReportsForm(forms.Form):
             start = timezone.make_aware(datetime(year, 1, 1))
         else:
             # Validate custom period
-            if self.cleaned_data['start_date']:
+            if not self.cleaned_data['start_date']:
                 raise ValidationError({'start_date': _('Missing date!')})
             if not self.cleaned_data['end_date']:
                 raise ValidationError({'end_date': _('Missing date!')})