+++ weblate/trans/models/unit.py
@@ -613,7 +613,7 @@ class Unit(models.Model, LoggerMixin):
                 report_error(error, sys.exc_info(), request)
                 change = self.change_set.all().order_by('-timestamp')[0]
             if change.author_id != request.user.id:
-                self.translation.commit_pending(request)
+                self.translation.commit_pending('pending unit', request)
 
         # Return if there was no change
         # We have to explicitly check for fuzzy flag change on monolingual
@@ -646,9 +646,9 @@ class Unit(models.Model, LoggerMixin):
         # Save updated unit to database
         self.save(backend=True)
 
-        old_translated = self.translation.stats.translated
-
         if change_action not in (Change.ACTION_UPLOAD, Change.ACTION_AUTO):
+            old_translated = self.translation.stats.translated
+
             # Update translation stats
             self.translation.invalidate_cache()
 
@@ -656,24 +656,24 @@ class Unit(models.Model, LoggerMixin):
             user.profile.translated += 1
             user.profile.save()
 
+            # Force commiting on completing translation
+            translated = self.translation.stats.translated
+            if (old_translated < translated and
+                    translated == self.translation.stats.all):
+                Change.objects.create(
+                    translation=self.translation,
+                    action=Change.ACTION_COMPLETE,
+                    user=user,
+                    author=user
+                )
+                self.translation.commit_pending('completed', request)
+
         # Notify subscribed users about new translation
         from weblate.accounts.notifications import notify_new_translation
         notify_new_translation(self, self.old_unit, user)
 
         # Generate Change object for this change
         self.generate_change(request, user, change_action)
-
-        # Force commiting on completing translation
-        translated = self.translation.stats.translated
-        if (old_translated < translated and
-                translated == self.translation.stats.all):
-            Change.objects.create(
-                translation=self.translation,
-                action=Change.ACTION_COMPLETE,
-                user=user,
-                author=user
-            )
-            self.translation.commit_pending(request)
 
         # Update related source strings if working on a template
         if self.translation.is_template: