+++ weblate/trans/views/basic.py
@@ -36,6 +36,7 @@ from django.utils.translation.trans_real import parse_accept_lang_header
 import django.views.defaults
 
 from weblate.checks.models import Check
+from weblate.formats.exporters import list_exporters
 from weblate.utils import messages
 from weblate.utils.stats import prefetch_stats
 from weblate.trans.models import (
@@ -56,6 +57,7 @@ from weblate.trans.views.helper import (
 )
 from weblate.trans.util import render, sort_objects, sort_unicode
 from weblate.vcs.gpg import get_gpg_public_key, get_gpg_sign_key
+from weblate.vcs.ssh import get_key_data
 
 
 def get_untranslated(base, limit=None):
@@ -487,6 +489,7 @@ def show_translation(request, project, component, lang):
                     pk=obj.pk
                 )
             ),
+            'exporters': list_exporters(),
         }
     )
 
@@ -542,8 +545,21 @@ def about(request):
             'title': _('About Weblate'),
             'versions': get_versions() + get_optional_versions(),
             'allow_index': True,
+        }
+    )
+
+
+def keys(request):
+    """Show keys information."""
+    return render(
+        request,
+        'keys.html',
+        {
+            'title': _('Weblate keys'),
             'gpg_key_id': get_gpg_sign_key(),
             'gpg_key': get_gpg_public_key(),
+            'ssh_key': get_key_data(),
+            'allow_index': True,
         }
     )