+++ weblate/trans/views/files.py
@@ -35,26 +35,24 @@ from weblate.trans.views.helper import (
 )
 
 
-def download_translation_format(request, project, lang, component=None):
+def download_translation(request, project, component, lang):
     obj = get_translation(request, project, component, lang)
 
-    form = DownloadForm(request.GET)
-    if not form.is_valid():
-        show_form_errors(request, form)
-        return redirect(obj)
+    kwargs = {}
 
-    units = obj.unit_set.search(
-        form.cleaned_data,
-        translation=obj,
-    )
+    if 'format' in request.GET or 'type' in request.GET:
+        form = DownloadForm(request.GET)
+        if not form.is_valid():
+            show_form_errors(request, form)
+            return redirect(obj)
 
-    return download_translation_file(obj, form.cleaned_data['format'], units)
+        kwargs['units'] = obj.unit_set.search(
+            form.cleaned_data,
+            translation=obj,
+        )
+        kwargs['fmt'] = form.cleaned_data['format']
 
-
-def download_translation(request, lang, project, component = ""):
-    obj = get_translation(request, project, component, lang)
-
-    return download_translation_file(obj)
+    return download_translation_file(obj, **kwargs)
 
 
 @require_POST